+++ src/pyopf/pointcloud/utils.py
@@ -1,3 +1,4 @@
+import os
 from pathlib import Path
 from urllib import parse
 
@@ -270,5 +271,6 @@ def add_buffers(gltf: pygltflib.GLTF2, buffers: dict[Path, Buffer], base_path: P
     for buffer in buffers.values():
         if buffer.filepath is None:
             raise ValueError("The buffer is not a memory mapped file")
-        uri = parse.quote(str(Path(buffer.filepath).relative_to(base_path)))
+        relpath = os.path.relpath(buffer.filepath, base_path)
+        uri = parse.quote(str(Path(relpath)))
         gltf.buffers.append(pygltflib.Buffer(byteLength=len(buffer), uri=uri))