+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## 1.3.0
+
+### Added
+
+- Parameter to opf2nerf to produce Nerfstudio-ready outputs
+- Example script to compute the reprojection error of input GCPs in calibrated cameras
+
 ## 1.2.0
 
 ### Added